+++ PCbuild/pythoncore.vcxproj.filters
@@ -1280,6 +1280,9 @@
     <ClCompile Include="..\Objects\unicode_formatter.c">
       <Filter>Objects</Filter>
     </ClCompile>
+    <ClCompile Include="..\Objects\unicode_writer.c">
+      <Filter>Objects</Filter>
+    </ClCompile>
     <ClCompile Include="..\Objects\unicodeobject.c">
       <Filter>Objects</Filter>
     </ClCompile>
